requirements.txt
Django==4.1.3
psycopg2-binary==2.9.5
pillow==9.3.0
djangorestframework==3.14.0